+++ tests/test_writer.py
@@ -131,4 +131,60 @@ class DatasetWriterTest(unittest.TestCase, IOCollector):
             # We know this will raise, but our test coverage increases a little bit.
             writer.write_dataset(ds)
 
+    def test_append_with_input_decode_cf(self):
+
+        src_path_pat = 'src_{}.zarr'
+        dst_path = 'my.zarr'
+        self.add_path(dst_path)
+
+        writer = DatasetWriter(dst_path, output_overwrite=False, input_decode_cf=False)
+
+        n = 3
+        for i in range(0, n):
+            src_dataset = new_test_dataset(day=i + 1)
+            src_path = src_path_pat.format(i)
+            self.add_path(src_path)
+            src_dataset.to_zarr(src_path)
+            with xr.open_zarr(src_path, decode_cf=False) as src_dataset:
+                writer.write_dataset(src_dataset, append=i > 0)
+
+        self._assert_time_slices_ok(dst_path, src_path_pat, n)
+
+    # see also https://github.com/pydata/xarray/issues/4412
+    def test_append_with_input_decode_cf_xarray(self):
+
+        src_path_pat = 'src_{}.zarr'
+        dst_path = 'my.zarr'
+        self.add_path(dst_path)
+
+        n = 3
+        for i in range(0, n):
+            src_dataset = new_test_dataset(day=i + 1)
+            src_path = src_path_pat.format(i)
+            self.add_path(src_path)
+            src_dataset.to_zarr(src_path)
+            with xr.open_zarr(src_path, decode_cf=False) as src_dataset:
+                if i == 0:
+                    src_dataset.to_zarr(dst_path, mode='w-')
+                else:
+                    # Hack:
+                    src_dataset = xr.decode_cf(src_dataset)
+                    for var_name in src_dataset.variables:
+                        src_dataset[var_name].encoding = {}
+                        src_dataset[var_name].attrs = {}
+                    src_dataset.to_zarr(dst_path, append_dim='time')
+
+        self._assert_time_slices_ok(dst_path, src_path_pat, n)
+
+    @classmethod
+    def _assert_time_slices_ok(cls, dst_path, src_path_pat, n):
+        with xr.open_zarr(dst_path, decode_cf=False) as ds:
+            for i in range(0, n):
+                src_path = src_path_pat.format(i)
+                with xr.open_zarr(src_path, decode_cf=False) as src_dataset:
+                    dst_dataset_slice = ds.isel(time=slice(i, i + 1)).compute()
+                    src_dataset.load()
+                    print(f'comparing time step {i} or {dst_path} with {src_path}')
+                    xr.testing.assert_allclose(dst_dataset_slice, src_dataset)
+
     # TODO: add real s3 tests using moto for boto mocking